Trouble shooting updates
Noticed a few issues when trying to validate the procedures against
some old data. Some minor changes to the handling of import names.
Seems to be fine but there are a few files for which I can’t figure out
the differences between dopOSCCI2 and 3… need to explore these in more
detail but also need to do some other work!