dopClip
added the ‘doClip’ function to cut the top of data to simulate what the
DWL software does if the limit is too low. I used this to check if it
was a problem and confirmed with Dorothy (Bishop) and OSCCI found that
it doesn’t significantly effect LI calculations. I’ll be reporting this
in a paper on infant data in the next couple of months - at least
that’s the plan