Remove event markers + behavioural update
Added a new aspect to the definition to allow for extra event markers
to be removed from the start or end of the file and fixed up some
graphing of the behavioural data. Actually, still something that needs
to happen with the naming of the plots when toggling but it’s getting
closer.