Updates: MATLAB 2020/1 plus event-marker short
Recent updates to MATLAB seem to have changed the iqr (inter-quartile range) function so I've written a dopIQR. There are a few updates to fit this in.
In January (doesn't look like I synced it) I also added a 'remove_short' setting to dopEventMarkers which removes event markers that are within the event_sep limit.